src/requirements.txt
langchain==0.2.7
langchain-community==0.2.7
anthropic==0.8.1
chromadb==0.4.22
streamlit==1.32.0
python-dotenv==1.0.0
sentence-transformers==2.5.1